Admission Prediction
The process of selecting candidates for admission to a graduate program
can be time-consuming and challenging for universities. To aid this process, we
aim to develop an Admission Prediction of Students model based on a dataset
that could include their scores of previous examinations, extracurricular
activities, letter of recommendation, research, etc.
Overall, the Admission Prediction of Students model will be a valuable
tool for universities to help identify the most qualified candidates for admission,
ultimately leading to better student outcomes and a more efficient admissions
process.